(modify:初始化加入电机响应等待)

+++ application/hand_task.h
@@ -48,7 +48,7 @@ typedef struct{
 
   /*feedback value*/
   fp32 feedback_motor_angle[HAND_MOTOR_COUNT];
-  int32_t feedback_motor_speed[HAND_MOTOR_COUNT];
+  fp32 feedback_motor_speed[HAND_MOTOR_COUNT];
   int32_t feedback_motor_current[HAND_MOTOR_COUNT];
 
   /*output value*/
